Les collisions fonctionnent et ca tient une ligne !

+++ DonjonGit/Copy of Donjon/src/fr/donjon/editor/EditorWindow.java
@@ -179,7 +179,7 @@ public class EditorWindow extends JFrame{
 
 					@Override
 					public void onValidate(Vecteur v) {
-						panDessin.reinitialize(v.x, v.y); //Reinitialisation avec la nouvelle taille
+						panDessin.reinitialize((int)v.x, (int)v.y); //Reinitialisation avec la nouvelle taille
 					}
 
 					@Override
